........
  r1350 | ckulla | 2009-12-05 13:55:06 -0800 (Sat, 05 Dec 2009) | 1 line
  
  Fix trivial compiler warning
........
  r1351 | robertm | 2009-12-12 15:49:20 -0800 (Sat, 12 Dec 2009) | 1 line
  
  iconvert: fixed error message when writing scanline to output file fail
........
  r1365 | lg | 2010-01-05 22:33:15 -0800 (Tue, 05 Jan 2010) | 3 lines
  
  Flush stdout after certain maketx status messages, so that in the middle of
  a long maketx execution, a calling process will get them right away.
........
  r1368 | lg | 2010-01-07 16:42:27 -0800 (Thu, 07 Jan 2010) | 3 lines
  
  Fixes from Henri Fousse: fix some TBB-related macros that were failing
  when compiling on Windows but not using TBB (mostly #if versus #ifdef issues).
........
  r1384 | wormszer | 2010-01-30 16:46:01 -0800 (Sat, 30 Jan 2010) | 3 lines
  
  Windows fixes: fix warnings and add math functions.
  Patch by: wormszer
........
  r1386 | lg | 2010-01-31 20:18:46 -0800 (Sun, 31 Jan 2010) | 3 lines
  
  Windows fix: when including windows.h, use the extra macros to limit the 
  stupid global defines it will introduce.
........
  r1393 | ckulla | 2010-02-08 17:40:32 -0800 (Mon, 08 Feb 2010) | 1 line
  
  Add portable sincos (reviewed as part of http://codereview.appspot.com/205045/show)
........
  r1394 | ckulla | 2010-02-08 17:52:37 -0800 (Mon, 08 Feb 2010) | 1 line
  
  Fix typo in previous commit
........


git-svn-id: http://svn.openimageio.org/oiio/branches/RB-0.7@1401 31904243-a364-42ac-a3db-9305396518ba

+++ src/fits.imageio/fitsoutput.cpp
@@ -172,7 +172,7 @@ FitsOutput::create_fits_header (void)
         std::vector<std::string> values;
         if (keyname == "Comment" || keyname == "History" || keyname == "Hierarch") {
             pystring::split (value, values, m_sep);
-            for (int i = 0; i < values.size(); ++i)
+            for (size_t i = 0; i < values.size(); ++i)
                 header += create_card (keyname, values[i]);
             continue;
         }
